backend/requirements.txt
alembic==1.14.1
aniso8601==10.0.0
attrs==25.1.0
blinker==1.9.0
click==8.1.8
extended==0.0.8
Flask==3.1.0
Flask-JWT==0.3.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
referencing==0.36.2
rpds-py==0.22.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
